* bootloader: (35 commits)
  term flash fix, + edge macros, encm error fix
  boot loader version info
  term fix, rene fail fix
  display HSE
  reordered comps
  fix cauto links
  save name of last failed link/set
  include all comps
  comp update to new init func
  mist config
  term width fix, autoconnect
  hal help + cons, save, load
  cons, term txqueue
  link cleanup
  + save/load, + update cmd/fb, link cleanup, main cleanup
  eeprom emulation working
  ld scripts don't like single line comments
  sector annotation in linker script
  hal sort fix
  hal sort fix
  ...

+++ term/basicdrawpane.hpp
@@ -27,5 +27,5 @@ public:
 private:
     wxLongLong time;
     std::vector<std::vector<float>> data;
-    int x,y,xpos,xstep;
+    double x,y,xpos,xstep;
 };
